switch to fastapi
