+++ templates/index.html
@@ -2,7 +2,7 @@
 <html>
 <head>
     <title>Execute PDF generation</title>
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
+    <link rel="stylesheet" type="text/css" href="/static/css/style.css">
 </head>
 <body>
     <div class="container">
